requirements.txt
setuptools
tqdm
nmslib
matplotlib
scikit-learn
Pillow
imgaug
imutils
easydict
pickle
tensorflow==1.14.0
numpy==1.15.4
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16